pyproject.toml
[tool.poetry]
name = "worder"
version = "0.1.0"
description = ""
authors = ["mefanya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.2"
sqlalchemy = "^2.0.41"
apscheduler = "^3.11.0"
python-dotenv = "^1.1.1"
alembic = "^1.16.2"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
